+++ protocols/kad/src/behaviour.rs
@@ -228,16 +228,6 @@ impl KademliaConfig {
         self
     }
 
-    /// Sets a custom protocol name.
-    ///
-    /// Kademlia nodes only communicate with other nodes using the same protocol
-    /// name. Using a custom name therefore allows to segregate the DHT from
-    /// others, if that is desired.
-    #[deprecated(since = "0.40.0", note = "use `set_protocol_names()` instead")]
-    pub fn set_protocol_name(&mut self, name: impl Into<Cow<'static, [u8]>>) -> &mut Self {
-        self.set_protocol_names(std::iter::once(name.into()).collect())
-    }
-
     /// Sets the timeout for a single query.
     ///
     /// > **Note**: A single query usually comprises at least as many requests