+++ CHANGELOG.md
@@ -43,6 +43,20 @@
 
 # `libp2p` facade crate
 
+# 0.49.0 - [unreleased]
+
+- Update to [`libp2p-tcp` `v0.37.0`](transports/tcp/CHANGELOG.md#0370).
+- 
+- Update to [`libp2p-noise` `v0.39.1`](transports/noise/CHANGELOG.md#0391).
+
+- Update to [`libp2p-swarm-derive` `v0.30.1`](swarm-derive/CHANGELOG.md#0301).
+
+- Update to [`libp2p-metrics` `v0.10.0`](misc/metrics/CHANGELOG.md#0100).
+
+- Update to [`libp2p-kad` `v0.41.0`](protocols/kad/CHANGELOG.md#0410).
+
+- Update to [`libp2p-noise` `v0.39.1`](transports/noise/CHANGELOG.md#0391).
+
 # 0.48.0
 
 - Update to [`libp2p-core` `v0.36.0`](core/CHANGELOG.md#0360).